+++ changelog.md
@@ -4,6 +4,8 @@
 
 ### Version 5.0.0
 
+* `cordapp`: Compatibility fix for Gradle 5.2 and above.
+
 * `api-scanner`: Compatibility fix for Gradle 5.x.
 
 * `cordformation`: Unload `SecurityProvider` instances installed by Network Bootstrapper to prevent linkage exceptions later.